Struct serde_protobuf::descriptor::Descriptors

pub struct Descriptors {
    // some fields omitted
}

A registry for any number of protocol buffer descriptors.

Methods

impl Descriptors

fn new() -> Descriptors

Creates a new empty descriptor set.

fn from_proto(file_set_proto: &FileDescriptorSet) -> Descriptors

Builds a descriptor set from the specified protocol buffer file descriptor set.

fn message_by_name(&self, name: &str) -> Option<&MessageDescriptor>

Looks up a message by its fully qualified name (i.e. .foo.package.Message).

fn enum_by_name(&self, name: &str) -> Option<&EnumDescriptor>

Looks up an enum by its fully qualified name (i.e. .foo.package.Enum).

fn add_file_set_proto(&mut self, file_set_proto: &FileDescriptorSet)

Adds all types defined in the specified protocol buffer file descriptor set to this registry.

fn add_file_proto(&mut self, file_proto: &FileDescriptorProto)

Adds all types defined in the specified protocol buffer file descriptor to this registry.

fn add_message_proto(&mut self, path: &str, message_proto: &DescriptorProto)

Adds a message and all nested types within that message from the specified protocol buffer descriptor.

fn add_message(&mut self, descriptor: MessageDescriptor)

Adds a single custom built message descriptor.

fn add_enum(&mut self, descriptor: EnumDescriptor)

Adds a single custom built enum descriptor.

fn resolve_refs(&mut self)

Resolves all internal descriptor type references, making them cheaper to follow.